+++ src/i18n/locales/en.ts
@@ -814,7 +814,6 @@ export default {
     "lattice": "Open play queue",
     "latticeLabel": "Play queue",
     "latticeBack": "Back",
-    "latticeSubtitle": "CURRENT PLAY QUEUE",
     "latticePlay": "PLAY",
     "latticeFocusCurrent": "Focus current song",
     "latticePlaybackControls": "Playback controls",
@@ -1980,7 +1979,6 @@ export default {
     "importGroupBackground": "Background",
     "importGroupSongTheme": "Song theme automation",
     "importGroupTrackCard": "Track card",
-    "importGroupLattice": "Queue collage",
     "importDerivedUnpin": "This will also turn off \"prefer custom theme\"",
     "importDerivedUploadedFont": "Your uploaded font will be replaced and deleted — you would need to upload it again",
     "importFontUnavailable": "This font is not installed here, so it will fall back to the builtin stack",
