+++ src/i18n/locales/in.ts
@@ -811,7 +811,6 @@ export default {
     "lattice": "Buka antrean putar",
     "latticeLabel": "Antrean putar",
     "latticeBack": "Kembali",
-    "latticeSubtitle": "ANTREAN PUTAR SAAT INI",
     "latticePlay": "PUTAR",
     "latticeFocusCurrent": "Fokus lagu saat ini",
     "latticePlaybackControls": "Kontrol pemutaran",
@@ -1895,7 +1894,6 @@ export default {
     "importGroupBackground": "Latar Belakang",
     "importGroupSongTheme": "Otomatisasi tema lagu",
     "importGroupTrackCard": "Kartu lagu",
-    "importGroupLattice": "Kolase antrean",
     "importDerivedUnpin": "Ini juga akan mematikan \"utamakan tema kustom\"",
     "importDerivedUploadedFont": "Font yang Anda unggah akan diganti dan dihapus — Anda perlu mengunggahnya lagi",
     "importFontUnavailable": "Font ini tidak diinstal di sini, jadi akan kembali ke tumpukan bawaan",
