+++ src/i18n/locales/zh-CN.ts
@@ -813,7 +813,6 @@ export default {
     "lattice": "打开播放队列",
     "latticeLabel": "播放队列",
     "latticeBack": "返回",
-    "latticeSubtitle": "当前播放队列",
     "latticePlay": "播放",
     "latticeFocusCurrent": "聚焦当前歌曲",
     "latticePlaybackControls": "播放控制",
@@ -1982,7 +1981,6 @@ export default {
     "importGroupBackground": "背景",
     "importGroupSongTheme": "歌曲主题自动化",
     "importGroupTrackCard": "歌曲信息卡片",
-    "importGroupLattice": "队列拼贴",
     "importDerivedUnpin": "将同时关闭「固定自定义主题」",
     "importDerivedUploadedFont": "你上传的字体将被替换并删除，需要重新上传",
     "importFontUnavailable": "本机未安装此字体，导入后会回落到内置字体栈",
